// src/pages/heavyMachine/roadRoller/types.d.ts
export interface RoadRollerContent {
  banner: ImageCMS;
  firstAbout: ImageCMS;
  secondAbout: ImageCMS;
  thirdAbout: ImageCMS;
  carousel: TextCMS;
  information: ImageCMS;
  tabs: ImageCMS;
  secondInformation: TextCMS;
  secondBanner: ImageCMS;
  attachments: TextCMS;
  platformRental: TextCMS;
}